refactor(tests): de-barrel tests/e2e/data/ per AGENTS.md
Last of the three de-barrel commits. Drop the data barrel and update
every importer (8 specs + connectedWallet fixture + 3 POMs that
import data constants) to import each constant directly from its
own file (./data/urls, ./data/settingsMenu, ./data/urlParams,
./data/themes).

Closes mmioana's PR comment on barrels. All three documented
'test-only exceptions' from tests/CLAUDE.md are now eliminated — the
suite aligns with AGENTS.md's unconditional no-barrels rule.
