--- tests/e2e/data/index.ts
@@ -1,4 +1,0 @@
-export * from './settingsMenu';
-export * from './themes';
-export * from './urlParams';
-export * from './urls';
